+++ src/veyeslack/commands/help.clj
@@ -12,6 +12,11 @@
                      "/veye list versionEye team1"]}
    :project {:text "shows a list of outdated project dependencies"
              :examples ["/veye project 52126476632bac1ae3007c88"]}
+   :cve {:text "shows a list of CVE alerts for the language/platform. \n
+               supported languages: ruby, nodejs, php, java, python, clojure."
+         :examples ["/veye cve ruby"
+                    "/veye cve nodejs 2"
+                    "/veye cve php"]}
    :help {:text "shows short list of commands or detailed help for the command"
           :examples ["/veye help"
                      "/veye help connect"]}