+++ src/veyeslack/services/versioneye.clj
@@ -1,12 +1,23 @@
 (ns veyeslack.services.versioneye
   (:require [clj-http.client :as http]
-            [clojurewerkz.spyglass.client :as spyglass]))
+            [clojure.string :as string]))
 
 (def api-url "https://www.versioneye.com/api")
 
 (defn to-api-uri
   [& path-items]
   (apply str (cons api-url path-items)))
+
+(defn to-language [lang]
+  (case (-> lang str string/lower-case)
+    "node" "Node.JS"
+    "nodejs" "Node.JS"
+    "js" "Node.JS"
+    "javascript" "Node.JS"
+    "clojure" "Java"
+    "php" "PHP"
+    "" "Java"
+    (string/capitalize lang)))
 
 (defn search
   [api-key term {:keys [n page] :or {n 10 page 1}}]
@@ -43,3 +54,12 @@
             {:accept :json
              :as :json
              :query-params {:api_key api-key}}))
+
+(defn security
+  [api-key {:keys [lang page] :or {lang "java" page 1}}]
+  (http/get (to-api-uri "/v2/security")
+            {:accept :json
+             :as :json
+             :query-params {:api-key api-key
+                            :language (to-language lang)
+                            :page (or page 1)}}))